Factor out arch-specific stuff into submodules of `libc`
